Move Operation to base and format_bytes32_string to utils
