+++ roles_royce/generic_method.py
@@ -4,7 +4,7 @@ from typing import Protocol
 from web3 import Web3
 from web3.types import HexStr
 
-from .roles_modifier import Operation
+from .protocols.base import Operation
 
 
 class Transactable(Protocol):
